+++ docs/conf.py
@@ -32,6 +32,7 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.intersphinx",
     "sphinx.ext.viewcode",
+    "sphinx_click",
 ]
 
 # Add any paths that contain templates here, relative to this directory.